gemfile: fix site building on newer ruby versions
csv, base64 and bigdecimal are no longer part of Ruby 3's standard library.
This change happened with Ruby 3.4.0, as such, the site cannot build on
newer Ruby versions without specifying these gems in the Gemfile.
Logger will be removed in 3.5.0, the last working version for this Jekyll
seems to be 1.6.6 so I've also added that to the Gemfile.